localedef: build errors with clang
/code/illumos-gate/usr/src/lib/libgen/common/mkdirp.c:184:31: error: comparison of integers of different signs: 'int' and 'size_t' (aka 'unsigned int') [-Werror,-Wsign-compare]
        for (wptr = wcPath, i = 0; i < wcPathlen; i++) {
                                   ~ ^ ~~~~~~~~~
1 error generated.

/code/illumos-gate/usr/src/cmd/localedef/wide.c:121:15: error: missing field 'nbits' initializer [-Werror,-Wmissing-field-initializers]
        { NULL, NULL },
                     ^
/code/illumos-gate/usr/src/cmd/localedef/wide.c:165:50: error: unused parameter 'n' [-Werror,-Wunused-parameter]
towide_none(wchar_t *c, const char *mb, unsigned n)
                                                 ^
/code/illumos-gate/usr/src/cmd/localedef/wide.c:227:9: error: comparison of integers of different signs: 'int' and 'unsigned int' [-Werror,-Wsign-compare]
        if (nb > n) {
            ~~ ^ ~
/code/illumos-gate/usr/src/cmd/localedef/wide.c:496:8: error: comparison of integers of different signs: 'unsigned int' and 'int' [-Werror,-Wsign-compare]
        if (n < width) {
            ~ ^ ~~~~~
4 errors generated.

/code/illumos-gate/usr/src/cmd/localedef/scanner.c:449:17: error: comparison of integers of different signs: 'int' and 'unsigned int' [-Werror,-Wsign-compare]
        if (mb_cur_max >= sizeof (mbs)) {
            ~~~~~~~~~~ ^  ~~~~~~~~~~~~
1 error generated.

/code/illumos-gate/usr/src/cmd/localedef/charmap.c:274:30: error: comparison of integers of different signs: 'size_t' (aka 'unsigned int') and 'int' [-Werror,-Wsign-compare]
            (strspn(e + si, digits) != (le - si)) ||
             ~~~~~~~~~~~~~~~~~~~~~~ ^   ~~~~~~~
/code/illumos-gate/usr/src/cmd/localedef/charmap.c:273:30: error: comparison of integers of different signs: 'size_t' (aka 'unsigned int') and 'int' [-Werror,-Wsign-compare]
            (strspn(s + si, digits) != (ls - si)) ||
             ~~~~~~~~~~~~~~~~~~~~~~ ^   ~~~~~~~
2 errors generated.

+++ usr/src/cmd/localedef/charmap.c
@@ -270,8 +270,8 @@ add_charmap_range(char *s, char *e, int wc)
 
 	if (((si = strcspn(s, digits)) == 0) || (si == ls) ||
 	    (strncmp(s, e, si) != 0) ||
-	    (strspn(s + si, digits) != (ls - si)) ||
-	    (strspn(e + si, digits) != (le - si)) ||
+	    (strspn(s + si, digits) != (size_t)(ls - si)) ||
+	    (strspn(e + si, digits) != (size_t)(le - si)) ||
 	    ((sn = atoi(s + si)) > ((en = atoi(e + si))))) {
 		errf(_("malformed charmap range"));
 		return;
